+++ storage/xml_test.go
@@ -3,8 +3,8 @@ package storage_test
 import (
 	"bytes"
 	"fmt"
-	"github.com/tomwright/dasel"
-	"github.com/tomwright/dasel/storage"
+	"github.com/tomwright/dasel/v2"
+	"github.com/tomwright/dasel/v2/storage"
 	"io"
 	"reflect"
 	"testing"